data integrity

+++ src/api/resolvers/categoryResolver.ts
@@ -2,6 +2,8 @@ import {GraphQLError} from 'graphql';
 import {Category, Species} from '../../types/DBTypes';
 import categoryModel from '../models/categoryModel';
 import {MyContext} from '../../types/MyContext';
+import speciesModel from '../models/speciesModel';
+import animalModel from '../models/animalModel';
 
 export default {
   Species: {
@@ -70,7 +72,24 @@ export default {
     deleteCategory: async (
       _parent: undefined,
       args: {id: string},
+      context: MyContext,
     ): Promise<{message: string; category?: Category}> => {
+      if (!context.userdata || context.userdata.role !== 'admin') {
+        throw new GraphQLError('User not authorized', {
+          extensions: {
+            code: 'UNAUTHORIZED',
+          },
+        });
+      }
+
+      // delete species and animals that belong to this category
+      const allSpecies = await speciesModel.find({category: args.id});
+      for (const species of allSpecies) {
+        await animalModel.deleteMany({species: species._id});
+      }
+
+      await speciesModel.deleteMany({category: args.id});
+
       const category = await categoryModel.findByIdAndDelete(args.id);
       if (category) {
         return {message: 'Category deleted', category};
